Allow JSON encoding as well.
Due to the asynchronicity of multithreaded connections we have
to use autodetection: by the time the "switch to json" command
has been processed the READ-MESSAGE loop has already been
invoked again, and so passing the decoder in as argument would
just give the old (stale) value again.

Other ideas discussed on #slime.